# docs/swagger.yaml
basePath: /
definitions:
  controller.Developer:
    properties:
      _id:
        example: "123456789"
        type: string
      bio:
        example: tech otaku save the world
        type: string
      name:
        example: HoYoverse
        type: string
    type: object
  controller.Games:
    properties:
      _id:
        example: "123456789"
        type: string
      desc:
        example: A tactical first-person shooter game developed by Riot Games
        type: string
      dev_name:
        $ref: '#/definitions/controller.Developer'
      game_banner:
        example: https://i.ibb.co.com/k1KdV7t/genshin-main-banner.png
        type: string
      game_logo:
        example: https://i.ibb.co.com/Z6xFZP6/genshin-logo.png
        type: string
      genre:
        example:
        - Adventure Games
        - Shooter
        - Action
        items:
          type: string
        type: array
      link_games:
        example: https://genshin.hoyoverse.com/id/
        type: string
      name:
        example: Valorant
        type: string
      preview:
        example: https://www.youtube.com/watch?v=qqnEjmnitgc
        type: string
      rating:
        example: 123.12
        type: number
      release_date:
        example: "2024-09-01T00:00:00Z"
        format: date-time
        type: string
    type: object
  controller.ReqDeveloper:
    properties:
      bio:
        example: tech otaku save the world
        type: string
      name:
        example: HoYoverse
        type: string
    type: object
  controller.ReqGames:
    properties:
      desc:
        example: A tactical first-person shooter game developed by Riot Games
        type: string
      dev_name:
        $ref: '#/definitions/controller.ReqDeveloper'
      game_banner:
        example: https://i.ibb.co.com/k1KdV7t/genshin-main-banner.png
        type: string
      game_logo:
        example: https://i.ibb.co.com/Z6xFZP6/genshin-logo.png
        type: string
      genre:
        example:
        - Adventure Games
        - Shooter
        - Action
        items:
          type: string
        type: array
      link_games:
        example: https://genshin.hoyoverse.com/id/
        type: string
      name:
        example: Valorant
        type: string
      preview:
        example: https://www.youtube.com/watch?v=qqnEjmnitgc
        type: string
      rating:
        example: 123.12
        type: number
    type: object
host: zenversegames-ba223a40f69e.herokuapp.com
info:
  contact:
    email: 714220044.@std.ulbi.ac.id
    name: API Support
    url: https://github.com/rayfanaqbil
  description: This is a sample swagger for Fiber
  title: TES SWAGGER ULBI
  version: "1.0"
paths:
  /games:
    get:
      consumes:
      - application/json
      description: Mengambil semua data games.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Games'
      summary: Get All Data Games.
      tags:
      - Games
  /games/{id}:
    get:
      consumes:
      - application/json
      description: Ambil per ID data games.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Games'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get By ID Data Games.
      tags:
      - Games
  /insert:
    post:
      consumes:
      - application/json
      description: Input data games.
      parameters:
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ReqGames'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Games'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Insert data Games.
      tags:
      - Games
schemes:
- https
- http
swagger: "2.0"